Display System Code in the Health Check (#11)
* Display System Code in the Health Check

System code appears in the formatted output when present in the json and is suppressed when not present. System code appears in brackets after the service name and is hyperlinked to the CMDB.

* Update health.html

Updated following feedback. System code presence now prefixes the Service Name with the Dewey logo hyperlinked to system code in Dewey.